protocol tool: refactoring
